requirements.txt
openai==1.3.0
python-dotenv==1.0.0
requests==2.31.0
prompt-toolkit==3.0.39
duckduckgo-search==3.9.3
